+++ nitrocli/CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 - Removed `storage status` subcommand
   - Moved its output into `status` command
 - Removed previously deprecated `--ascii` option from `otp set` command
+- Bumped `nitrokey` dependency to `0.4.0-alpha.0`
 
 
 0.2.4